input required new trick
