+++ src/Form/TricksType.php
@@ -21,8 +21,12 @@ class TricksType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('title')
-            ->add('content')
+            ->add('title',null,[
+            'required' => true,
+                ])
+            ->add('content',null,[
+                'required' => true,
+            ])
             ->add('images', FileType::class, [
                 'label' => 'Images',
                 'multiple' => true,
@@ -56,12 +60,13 @@ class TricksType extends AbstractType
         'label' => 'Image',
 
 
+
         // unmapped means that this field is not associated to any entity property
         'mapped' => false,
 
         // make it optional so you don't have to re-upload the PDF file
         // every time you edit the Product details
-        'required' => false,
+        'required' => true,
 
         // unmapped fields can't define their validation using annotations
         // in the associated entity, so you can use the PHP constraint classes
