+++ templates/tricks/_form.html.twig
@@ -11,7 +11,7 @@
 
 
                 <div class= " col-md-6">
-                    {{ form_row(form.image ,{'id': 'imgInp' , 'label' : 'Image', 'attr' : { 'class' : 'form-control-file' }} ) }}
+                    {{ form_row(form.image ,{'id': 'imgInp' , 'label' : 'Image', 'attr' : { 'class' : 'form-control-file'}} ) }}
                 </div>
